move to OOP approach, this is pretty now
