+++ src/pages/events.astro
@@ -2,7 +2,7 @@
 import Layout from '../layouts/Layout.astro';
 import Event from '../components/Events.svelte';
 import Header from '../components/Header.svelte';
-import { normalWrapper } from '../lib/db';
+import { normalWrapper } from '../lib/backend';
 
 let session = { user: { username: 'DEV', emailVerified: true, user_role: 'admin', user_team: 'XXX' } };
