+++ src/pages/events/[uuid].astro
@@ -2,7 +2,7 @@
 import Layout from '../../layouts/Layout.astro';
 import Challenge from '../../components/Challenges.svelte';
 import Header from '../../components/Header.svelte';
-import { normalWrapper } from '../../lib/db';
+import { normalWrapper } from '../../lib/backend';
 
 const { uuid } = Astro.params;
